Report Functionality
Backend working to generate a report for date range or for all sessions in a roster.  Page is currently hardcoded to do report for all sessions while awaiting front end.

DatabaseUtilities: session list returned is sorted.  Bugfix on syntax error.

professormain: if a session was chosen for editing, clears it because this is the page where you can change rosters.